+++ Launcher/FF8.cs
@@ -1,4 +1,7 @@
-namespace F8Framework.Core
+using F8Framework.Core;
+using F8Framework.F8ExcelDataClass;
+
+namespace F8Framework.Launcher
 {
     public static class FF8
     {