+++ Tests/FSM/DemoFSM.cs
@@ -1,5 +1,6 @@
 ﻿using UnityEngine;
 using F8Framework.Core;
+using F8Framework.Launcher;
 
 namespace F8Framework.Tests
 {