+++ Tests/Network/MultiNetworkChannel/MultiNetworkServer.cs
@@ -2,6 +2,7 @@
 using UnityEngine;
 using UnityEngine.UI;
 using F8Framework.Core;
+using F8Framework.Launcher;
 
 namespace F8Framework.Tests
 {