+++ Tests/UI/DemoUIManager.cs
@@ -1,5 +1,6 @@
 using System.Collections.Generic;
 using F8Framework.Core;
+using F8Framework.Launcher;
 using UnityEngine;
 
 namespace F8Framework.Tests